Move some dependencies around
