+++ apis/cloudflare/src/metric-aggregator.ts
@@ -1,14 +1,7 @@
-import {
-  DataPoint,
-  DataPointType,
-  Histogram,
-  MetricData,
-  ResourceMetrics,
-} from "@opentelemetry/sdk-metrics";
-import { HrTime } from "@opentelemetry/api";
-import { hrTimeToMicroseconds } from "@opentelemetry/core";
+import { MetricData, ResourceMetrics } from "@opentelemetry/sdk-metrics";
 import { Resource } from "@opentelemetry/resources";
-import { PrometheusSerializer } from "./PrometheusSerializer";
+import { PrometheusSerializer } from "@braintrust/proxy/src/PrometheusSerializer";
+import { aggregateMetrics, prometheusSerialize } from "@braintrust/proxy";
 
 declare global {
   interface Env {
@@ -17,12 +10,13 @@ declare global {
     // of other distinct sets of labels) works out to one Prometheus timeseries. Shards allow us to
     // essentially aggregate _across_ workers.
     METRICS_SHARDS?: number;
-    // If a metric doesn't show up for this many seconds, it'll be deleted from the store. We detect
+    // TODO: If a metric doesn't show up for this many seconds, it'll be deleted from the store. We detect
     // this at read time.
     METRICS_TTL?: number;
   }
 }
 
+// XXX Move the meat of this implementation into the proxy package (with callback hooks to durable storage)
 export class PrometheusMetricAggregator {
   state: DurableObjectState;
   constructor(state: DurableObjectState, env: Env) {
@@ -47,69 +41,16 @@ export class PrometheusMetricAggregator {
 
   async handlePush(request: Request): Promise<Response> {
     const metrics = (await request.json()) as ResourceMetrics;
-
-    for (const scopeMetrics of metrics.scopeMetrics) {
-      for (const metric of scopeMetrics.metrics) {
-        for (let i = 0; i < metric.dataPoints.length; i++) {
-          // NOTE: We should be able to batch these get operations
-          // into sets of keys at most 128 in length
-          const metricKey =
-            "otel_metric_" +
-            JSON.stringify({
-              name: metric.descriptor.name,
-              dataPointType: metric.dataPointType,
-              labels: metric.dataPoints[i].attributes,
-            });
-
-          let existing = (await this.state.storage.get<MetricData>(
-            metricKey,
-          )) || {
-            ...metric,
-            dataPoints: [],
-          };
-          if (existing && existing.dataPointType !== metric.dataPointType) {
-            throw new Error("Invalid data point (type mismatch)");
-          }
-
-          let newValue = undefined;
-          switch (metric.dataPointType) {
-            case DataPointType.SUM:
-              newValue = coalesceFn(
-                existing.dataPoints[0] as DataPoint<number>,
-                metric.dataPoints[i],
-                mergeCounters,
-              );
-              break;
-            case DataPointType.GAUGE:
-              newValue = coalesceFn(
-                existing.dataPoints[0] as DataPoint<number>,
-                metric.dataPoints[i],
-                mergeGauges,
-              );
-              break;
-            case DataPointType.HISTOGRAM:
-              newValue = coalesceFn(
-                existing.dataPoints[0] as DataPoint<Histogram>,
-                metric.dataPoints[i],
-                mergeHistograms,
-              );
-              break;
-            case DataPointType.EXPONENTIAL_HISTOGRAM:
-              return new Response("Not Implemented: Exponential Histogram", {
-                status: 501,
-              });
-          }
-
-          if (newValue !== undefined) {
-            (existing as any).descriptor = metric.descriptor; // Update the descriptor in case the code changes it
-            existing.dataPoints[0] = newValue;
-            // See "Write buffer behavior" in https://developers.cloudflare.com/durable-objects/api/transactional-storage-api/
-            // The only reason to await this put is to apply backpressure, which should be unnecessary given the small # of metrics
-            // we're aggregating over
-            this.state.storage.put(metricKey, existing);
-          }
-        }
-      }
+    try {
+      await aggregateMetrics(
+        metrics,
+        async (key: string) =>
+          (await this.state.storage.get<MetricData>(key)) || null,
+        (key: string, value: MetricData) => this.state.storage.put(key, value),
+      );
+    } catch (e) {
+      console.error("Error aggregating metrics", e);
+      return new Response("Error aggregating metrics", { status: 500 });
     }
 
     return new Response(null, { status: 204 });
@@ -144,9 +85,8 @@ export class PrometheusMetricAggregator {
         },
       ],
     };
-    const serializer = new PrometheusSerializer("", true /*appendTimestamp*/);
 
-    return new Response(serializer.serialize(resourceMetrics), {
+    return new Response(prometheusSerialize(resourceMetrics), {
       headers: {
         "Content-Type": "text/plain",
       },
@@ -161,82 +101,4 @@ export class PrometheusMetricAggregator {
   static metricsTTL(env: Env): number {
     return env.METRICS_TTL ?? 24 * 7 * 3600;
   }
-}
-
-function mergeHistograms(
-  base: DataPoint<Histogram>,
-  delta: DataPoint<Histogram>,
-): DataPoint<Histogram> {
-  if (
-    JSON.stringify(base.value.buckets.boundaries) !==
-    JSON.stringify(delta.value.buckets.boundaries)
-  ) {
-    throw new Error(
-      "Unsupported: merging histograms with different bucket boundaries",
-    );
-  }
-
-  return {
-    startTime: minHrTime(base.startTime, delta.startTime),
-    endTime: maxHrTime(base.endTime, delta.endTime),
-    attributes: { ...base.attributes } /* these are assumed to be the same */,
-    value: {
-      buckets: {
-        boundaries: [...base.value.buckets.boundaries],
-        counts: base.value.buckets.counts.map(
-          (count, i) => count + delta.value.buckets.counts[i],
-        ),
-      },
-      sum: (base.value.sum || 0) + (delta.value.sum || 0),
-      count: base.value.count + delta.value.count,
-      min: coalesceFn(base.value.max, delta.value.max, Math.min),
-      max: coalesceFn(base.value.max, delta.value.max, Math.max),
-    },
-  };
-}
-
-function mergeGauges(
-  base: DataPoint<number>,
-  delta: DataPoint<number>,
-): DataPoint<number> {
-  const baseT = hrTimeToMicroseconds(base.endTime);
-  const deltaT = hrTimeToMicroseconds(delta.endTime);
-  return {
-    startTime: deltaT >= baseT ? base.startTime : delta.startTime,
-    endTime: maxHrTime(base.endTime, delta.endTime),
-    attributes: { ...base.attributes } /* these are assumed to be the same */,
-    value: deltaT >= baseT ? delta.value : base.value,
-  };
-}
-
-function mergeCounters(
-  base: DataPoint<number>,
-  delta: DataPoint<number>,
-): DataPoint<number> {
-  return {
-    startTime: minHrTime(base.startTime, delta.startTime),
-    endTime: maxHrTime(base.endTime, delta.endTime),
-    attributes: { ...base.attributes } /* these are assumed to be the same */,
-    value: base.value + delta.value,
-  };
-}
-
-function minHrTime(a: HrTime, b: HrTime): HrTime {
-  const at = hrTimeToMicroseconds(a);
-  const bt = hrTimeToMicroseconds(b);
-  return at <= bt ? a : b;
-}
-
-function maxHrTime(a: HrTime, b: HrTime): HrTime {
-  const at = hrTimeToMicroseconds(a);
-  const bt = hrTimeToMicroseconds(b);
-  return at >= bt ? a : b;
-}
-
-function coalesceFn<T>(
-  a: T | undefined,
-  b: T | undefined,
-  coalesce: (a: T, b: T) => T,
-): T | undefined {
-  return a === undefined ? b : b === undefined ? a : coalesce(a, b);
 }
