+++ packages/proxy/package.json
@@ -68,6 +68,7 @@
   },
   "dependencies": {
     "@opentelemetry/api": "^1.7.0",
+    "@opentelemetry/core": "^1.18.1",
     "@opentelemetry/resources": "^1.18.1",
     "@opentelemetry/sdk-metrics": "^1.18.1",
     "eventsource-parser": "^1.1.1",
